chore: resolve todo batch 7 (013 reduced, 009 scoped)
- 013: remaining items from the perf bundle after 027/010 handled
  the rest.
  * Sort rules by strength at database load time via a stable
    sort_rules_by_strength_recursive helper in
    evaluator/strength.rs, called from
    MagicDatabase::with_builtin_rules_and_config and
    load_from_file_with_config. Stable sort preserves source order
    for equal-strength rules so no tests needed loosening.
  * Extract read_bytes_at const-generic helper in
    evaluator/types/mod.rs and use it from numeric.rs, float.rs,
    date.rs (six call sites), eliminating the duplicated bounds-
    check + try_into preamble.
- 009 (scoped to top offenders):
  * Split src/parser/grammar/mod.rs (1240 -> 796 lines) into three
    files: numbers.rs (number parsers), value.rs (hex/string/float/
    numeric value parsers), and mod.rs (grammar orchestration and
    tests). Public API unchanged via re-exports.
  * Extract EvaluationConfig and its constructors/validators from
    lib.rs (916 -> 624 lines) into a new src/config.rs module.
    libmagic_rs::EvaluationConfig still resolves via re-export so
    external consumers are unaffected.
  * output/json.rs is already <800 production lines (the bulk is
    tests), so it was left alone. Other files over 800 are still
    open for follow-up but are not in scope for this pass.
