+++ src/evaluator/types/date.rs
@@ -1,7 +1,7 @@
 // Copyright (c) 2025-2026 the libmagic-rs contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use super::TypeReadError;
+use super::{TypeReadError, read_bytes_at};
 use crate::parser::ast::{Endianness, Value};
 
 /// Day-of-week names matching GNU `file` output format.
@@ -49,17 +49,7 @@ pub fn read_date(
     endian: Endianness,
     utc: bool,
 ) -> Result<Value, TypeReadError> {
-    let end = offset.checked_add(4).ok_or(TypeReadError::BufferOverrun {
-        offset,
-        buffer_len: buffer.len(),
-    })?;
-    let arr: [u8; 4] = buffer
-        .get(offset..end)
-        .and_then(|s| s.try_into().ok())
-        .ok_or(TypeReadError::BufferOverrun {
-            offset,
-            buffer_len: buffer.len(),
-        })?;
+    let arr: [u8; 4] = read_bytes_at(buffer, offset)?;
 
     let secs = match endian {
         Endianness::Little => u32::from_le_bytes(arr),
@@ -107,17 +97,7 @@ pub fn read_qdate(
     endian: Endianness,
     utc: bool,
 ) -> Result<Value, TypeReadError> {
-    let end = offset.checked_add(8).ok_or(TypeReadError::BufferOverrun {
-        offset,
-        buffer_len: buffer.len(),
-    })?;
-    let arr: [u8; 8] = buffer
-        .get(offset..end)
-        .and_then(|s| s.try_into().ok())
-        .ok_or(TypeReadError::BufferOverrun {
-            offset,
-            buffer_len: buffer.len(),
-        })?;
+    let arr: [u8; 8] = read_bytes_at(buffer, offset)?;
 
     let secs = match endian {
         Endianness::Little => u64::from_le_bytes(arr),
